review/cleanup task-role
