+++ locales_dev/en/translation.json
@@ -115,7 +115,6 @@
   "Expires:": "Expires:",
   "Get Help": "Get Help",
   "GitHub Users": "GitHub Users",
-  "Github Collaborators": "Github Collaborators",
   "Go Back": "Go Back",
   "Heading": "Heading",
   "Hello! What can Metecho help you do today?": "Hello! What can Metecho help you do today?",
@@ -151,6 +150,7 @@
   "Next Steps for this Epic": "Next Steps for this Epic",
   "Next Steps for this Task": "Next Steps for this Task",
   "No matching branches found.": "No matching branches found.",
+  "No users found.": "No users found.",
   "Not Enabled": "Not Enabled",
   "Notify Assigned Developer by Email": "Notify Assigned Developer by Email",
   "Notify Assigned Tester by Email": "Notify Assigned Tester by Email",
@@ -161,7 +161,7 @@
   "Org Type": "Org Type",
   "Org Type cannot be changed while a scratch org exists for this task.": "Org Type cannot be changed while a scratch org exists for this task.",
   "Org:": "Org:",
-  "Other Github Collaborators": "Other Github Collaborators",
+  "Other GitHub Collaborators": "Other GitHub Collaborators",
   "Package Directories": "Package Directories",
   "Pending review": "Pending review",
   "Plan Walkthrough": "Plan Walkthrough",
@@ -178,8 +178,7 @@
   "Project Scratch Org": "Project Scratch Org",
   "Project:": "Project:",
   "Projects": "Projects",
-  "Quick Find": "Quick Find",
-  "Re-Sync Github Collaborators": "Re-Sync Github Collaborators",
+  "Re-Sync GitHub Collaborators": "Re-Sync GitHub Collaborators",
   "Re-Sync Projects": "Re-Sync Projects",
   "Reassigning Org Ownership…": "Reassigning Org Ownership…",
   "Refresh Org": "Refresh Org",
@@ -205,7 +204,7 @@
   "Save & Next": "Save & Next",
   "Saving Ignored Changes…": "Saving Ignored Changes…",
   "Saving…": "Saving…",
-  "Search": "Search",
+  "Search for user": "Search for user",
   "Select a Project": "Select a Project",
   "Select a box below to discover what’s possible.": "Select a box below to discover what’s possible.",
   "Select a branch to use for this epic": "Select a branch to use for this epic",
@@ -247,7 +246,7 @@
   "Successfully submitted epic for review on GitHub: “{{epic_name}}”.": "Successfully submitted epic for review on GitHub: “{{epic_name}}”.",
   "Successfully submitted review for task: “{{task_name}}”.": "Successfully submitted review for task: “{{task_name}}”.",
   "Successfully submitted task for testing: “{{task_name}}”.": "Successfully submitted task for testing: “{{task_name}}”.",
-  "Syncing Collaborators…": "Syncing Collaborators…",
+  "Syncing GitHub Collaborators…": "Syncing GitHub Collaborators…",
   "Syncing Projects…": "Syncing Projects…",
   "Task": "Task",
   "Task Name": "Task Name",
@@ -266,6 +265,8 @@
   "Tester": "Tester",
   "The existing Dev Org for this task has unretrieved changes. Changing the assigned developer will also delete the org, and any changes will be lost. Are you sure you want to do that?": "The existing Dev Org for this task has unretrieved changes. Changing the assigned developer will also delete the org, and any changes will be lost. Are you sure you want to do that?",
   "The existing Dev Org for this task has unretrieved changes. Removing the assigned developer will also delete the org, and any changes will be lost. Are you sure you want to do that?": "The existing Dev Org for this task has unretrieved changes. Removing the assigned developer will also delete the org, and any changes will be lost. Are you sure you want to do that?",
+  "There are no available Epic Collaborators.": "There are no available Epic Collaborators.",
+  "There are no available GitHub Collaborators.": "There are no available GitHub Collaborators.",
   "There aren't any available branches at this time.": "There aren't any available branches at this time.",
   "This becomes bold text": "This becomes bold text",
   "This becomes italic text": "This becomes italic text",
@@ -313,7 +314,7 @@
   "What is Metecho?": "What is Metecho?",
   "You are in offline mode. We are trying to reconnect, but you may need to": "You are in offline mode. We are trying to reconnect, but you may need to",
   "anErrorOccurred": "An error occurred. Try the <1>home page</1>?",
-  "assignUserHelper": "Assign any Github user to this role. If they are not already, the will also be added as an Epic Collaborator.",
+  "assignUserHelper": "Assign any user to this role, and they will also be added as an Epic Collaborator.",
   "avatar for user {{username}}": "avatar for user {{username}}",
   "check again": "check again",
   "confirmDeleteEpic": "Are you sure you want to delete epic “<1>{{name}}</1>”? This will also delete any tasks and scratch orgs in this epic.",
@@ -328,15 +329,13 @@
   "createScratchOrgHelp": "<0>You are creating a Scratch Org for <1>{{type}}</1> “<3>{{name}}</3>”.</0><1>Your new org will expire in 30 days.<1></1>You will be able to access your org from this <3>{{type}}</3> page.</1><2><0>You will not be able to retrieve any changes made in this Scratch Org.</0></2>",
   "devHubInfo": "Connection to a Salesforce org with Dev Hub enabled is required to create a Dev or Test scratch org. Learn how to <2>create a Developer Edition org</2> and <6>enable Dev Hub</6>.",
   "devHubNotEnabled": "This Salesforce org does not have Dev Hub enabled or your user does not have permission to create scratch orgs. Learn how to <2>enable Dev Hub</2>.",
-  "epicCollaborators": "Only users who have access to the GitHub repository for this epic will appear in the list below. Visit GitHub to invite additional collaborators.",
+  "epicCollaborators": "Only users who have access to the GitHub repository for this Epic will appear in the list below. Visit GitHub to invite additional collaborators.",
   "epicNotFound": "We can’t find the epic you’re looking for. Try <2>another epic</2> from that project?",
   "errorWithAccount": "An error occurred with your account. Try the <2>home page</2>?",
   "ignored": "ignored",
   "markdownGuide": "For more options, view this <2>Markdown Guide</2>.",
-  "noEpicCollaborators": "There are no Epic Collaborators. Select GitHub users to add as Collaborators.",
   "noGitHubUsers": "We couldn’t find any GitHub users who have access to this project. Try re-syncing the list of available collaborators, or contact an admin for this project on GitHub.",
   "noProjectsHelper": "We couldn’t find any projects you have access to on GitHub. Confirm that you are logged into the correct account or contact an admin on GitHub.",
-  "noUsersFound": "No users found.",
   "notifyUser": "<0>Notify</0><1></1> <3><0><0>{{username}}</0></0> by email</3>",
   "orgNotYetCreated": "<0>Status:</0> not yet created",
   "orgOwnedByWrongUser": "<0>Status:</0> owned by user",
