+++ src/js/store/user/reducer.ts
@@ -5,7 +5,7 @@ export interface GitHubUser {
   id: string;
   login: string;
   avatar_url: string;
-  name: string;
+  name?: string;
 }
 
 export interface User {
