+++ test/js/components/epics/detail.test.jsx
@@ -10,22 +10,18 @@ import {
   updateObject,
 } from '~js/store/actions';
 import { refreshGitHubUsers } from '~js/store/projects/actions';
-import { getUrlParam, removeUrlParam } from '~js/utils/api';
 import routes from '~js/utils/routes';
 
 import { renderWithRedux, storeWithThunk } from './../../utils';
 
 jest.mock('~js/store/actions');
 jest.mock('~js/store/projects/actions');
-jest.mock('~js/utils/api');
 
 fetchObject.mockReturnValue(() => Promise.resolve({ type: 'TEST' }));
 fetchObjects.mockReturnValue(() => Promise.resolve({ type: 'TEST' }));
 updateObject.mockReturnValue(() => Promise.resolve({ type: 'TEST' }));
 createObject.mockReturnValue(() => Promise.resolve({ type: 'TEST' }));
 refreshGitHubUsers.mockReturnValue(() => Promise.resolve({ type: 'TEST' }));
-getUrlParam.mockReturnValue(null);
-removeUrlParam.mockReturnValue('');
 
 afterEach(() => {
   fetchObject.mockClear();
@@ -33,8 +29,6 @@ afterEach(() => {
   updateObject.mockClear();
   createObject.mockClear();
   refreshGitHubUsers.mockClear();
-  getUrlParam.mockClear();
-  removeUrlParam.mockClear();
 });
 
 const defaultOrg = {
@@ -213,18 +207,14 @@ describe('<EpicDetail/>', () => {
     const opts = Object.assign({}, defaults, options);
     const { initialState, projectSlug, epicSlug } = opts;
     const context = {};
-    const history = { replace: jest.fn() };
     const response = renderWithRedux(
       <StaticRouter context={context}>
-        <EpicDetail
-          match={{ params: { projectSlug, epicSlug } }}
-          history={history}
-        />
+        <EpicDetail match={{ params: { projectSlug, epicSlug } }} />
       </StaticRouter>,
       initialState,
       storeWithThunk,
     );
-    return { ...response, context, history };
+    return { ...response, context };
   };
 
   test('renders epic detail, scratch org, and tasks list', () => {
@@ -249,23 +239,6 @@ describe('<EpicDetail/>', () => {
 
     expect(getByText('Add a Task for Epic 1')).toBeVisible();
     expect(queryByText('Tasks for Epic 1')).toBeNull();
-  });
-
-  describe('`SHOW_EPIC_COLLABORATORS` param is truthy', () => {
-    beforeAll(() => {
-      getUrlParam.mockReturnValue('true');
-    });
-
-    afterAll(() => {
-      getUrlParam.mockReturnValue(null);
-    });
-
-    test('opens assign-users modal', () => {
-      const { history, getByText } = setup();
-
-      expect(history.replace).toHaveBeenCalledWith({ search: '' });
-      expect(getByText('GitHub Users')).toBeVisible();
-    });
   });
 
   describe('epic not found', () => {
@@ -352,7 +325,7 @@ describe('<EpicDetail/>', () => {
     });
   });
 
-  describe('<AssignUsersModal />', () => {
+  describe('<AssignEpicCollaboratorsModal />', () => {
     test('opens and closes', () => {
       const { getByText, getByTitle, queryByText } = setup();
       fireEvent.click(getByText('Add or Remove Collaborators'));
