+++ test/js/components/orgs/cards.test.jsx
@@ -8,8 +8,6 @@ import TaskOrgCards, {
 } from '~js/components/orgs/taskOrgCards';
 import { deleteObject, updateObject } from '~js/store/actions';
 import { refetchOrg } from '~js/store/orgs/actions';
-import { addUrlParams } from '~js/utils/api';
-import { SHOW_EPIC_COLLABORATORS } from '~js/utils/constants';
 
 import {
   renderWithRedux,
@@ -245,25 +243,6 @@ describe('<TaskOrgCards/>', () => {
 
       expect(data.assigned_dev.login).toEqual('user-name');
       expect(data.should_alert_dev).toBe(true);
-    });
-
-    test('redirects to epic-detail if no users to assign', () => {
-      const task = {
-        ...defaultTask,
-        assigned_dev: null,
-      };
-      const epicUsers = [];
-      const { getByText, context } = setup({
-        task,
-        epicUsers,
-        assignUserModalOpen: 'Dev',
-      });
-      fireEvent.click(getByText('View Epic to Add Collaborators'));
-
-      expect(context.action).toEqual('PUSH');
-      expect(context.url).toEqual(
-        addUrlParams('epic-url', { [SHOW_EPIC_COLLABORATORS]: true }),
-      );
     });
   });
